Extracted all formatting from TextUI class to Formatter class, Formatter class now does the formatting of text for display
